:wrench: Add close action to hidden menu bar
