+++ src/main/app.ts
@@ -90,6 +90,55 @@ export class KandoApp {
   /** This contains the last IWMInfo which was received. */
   private lastWMInfo?: IWMInfo;
 
+  /**
+   * Most of the initialization is done in the init() method. This constructor is only
+   * used to set up the hidden menu bar as this has to be done before the Electron app is
+   * ready.
+   */
+  constructor() {
+    // On macOS, we loose the copy and paste functionality when using no menu bar. So we
+    // add a hidden menu bar with some default actions. We also add a custom handler for
+    // closing the window, so that we can hide the menu window instead of closing it.
+    const template = [
+      {
+        label: 'Edit',
+        submenu: [
+          {
+            role: 'undo',
+          },
+          {
+            role: 'redo',
+          },
+          {
+            role: 'cut',
+          },
+          {
+            role: 'copy',
+          },
+          {
+            role: 'paste',
+          },
+          {
+            label: 'Close',
+            accelerator: process.platform === 'darwin' ? 'Cmd+W' : 'Alt+F4',
+            click: () => {
+              console.log('Close menu window');
+              if (this.settingsWindow?.isFocused()) {
+                this.settingsWindow.close();
+              } else if (this.menuWindow?.isVisible()) {
+                this.menuWindow.hide();
+              }
+            },
+          },
+        ],
+      },
+    ];
+    const menu = Menu.buildFromTemplate(
+      template as Electron.MenuItemConstructorOptions[]
+    );
+    Menu.setApplicationMenu(menu);
+  }
+
   /** This is called when the app is started. It initializes the backend and the window. */
   public async init() {
     // Bail out if the backend is not available.
